Move files to more standard folder layout
