+++ lib/lightstep/tracer/thrift/reporting_service.rb
