+++ lib/lightstep/tracer/thrift/types.rb
