+++ lib/lightstep/tracer/transports/transport_http_json.rb
