+++ lib/lightstep/tracer/util.rb
